+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   annotated.
 - Fixed handling of inferred generic types
 - Added missing handling for property get/set accessors
+- String-like and number-like return types (enums mainly) now emit the appropriate type for design:returntype
 
 # v0.1.3
 - Fixes an additional case where `null` was used instead of `undefined` when generating typescript AST elements